moved the process from docapp.js to app.js and refactored it a bit
